+++ Project.toml
@@ -24,6 +24,9 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 TransformsBase = "28dd2a49-a57a-4bfb-84ca-1a49db9b96b8"
 
 [compat]
+LinearAlgebra="1.6"
+Random="1.6"
+Statistics="1.6"
 CategoricalArrays = "0.10"
 CategoricalDistributions = "0.1"
 Clustering = "0.15"
