### 2024-01-06 142b84f
🐛 Fix freezing dict in group_inds #73
