+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Imbalance"
 uuid = "c709b415-507b-45b7-9a3d-1767c89fde68"
 authors = ["Essam Wisam <essamwisam@outlook.com>", "Anthony Blaom <anthony.blaom@gmail.com> and contributors"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
