+++ v0.1.3/examples/colab.sh
@@ -1,0 +1,9 @@
+if ! command -v julia 3>&1 > /dev/null
+then
+    wget -q 'https://julialang-s3.julialang.org/bin/linux/x64/1.7/julia-1.7.2-linux-x86_64.tar.gz' \
+        -O /tmp/julia.tar.gz
+    tar -x -f /tmp/julia.tar.gz -C /usr/local --strip-components 1
+    rm /tmp/julia.tar.gz
+fi
+julia -e 'using Pkg; pkg"add IJulia; precompile;"'
+echo 'Done. Change the runtime to `Julia` by choosing `Runtime` from the toolbar then `Change runtime type`. You can then delete this cell.'
