### 2024-01-09 031af23
🚑 Fix scitypes for all existing methods
