+++ src/undersampling_methods/random_undersample/interface_mlj.jl
@@ -50,9 +50,20 @@ MMI.metadata_pkg(
 
 MMI.metadata_model(
     RandomUndersampler,
-    input_scitype = Union{Table(Continuous), AbstractMatrix{Continuous}},
-    output_scitype = Union{Table(Continuous), AbstractMatrix{Continuous}},
-    target_scitype = AbstractVector,
+    input_scitype = Tuple{
+                        Union{
+                            Table(Continuous),
+                            AbstractMatrix{Continuous}
+                        }, 
+                        AbstractVector
+                    },
+    output_scitype = Tuple{
+        Union{
+            Table(Continuous),
+            AbstractMatrix{Continuous}
+        }, 
+        AbstractVector
+    },
     load_path = "Imbalance.MLJ.RandomUndersampler"
 )
 function MMI.transform_scitype(s::RandomUndersampler)
