+++ src/oversampling_methods/smotenc/interface_mlj.jl
@@ -61,11 +61,12 @@ MMI.metadata_model(
     SMOTENC,
     input_scitype = Union{
         Table(Union{Infinite, Finite}),
+        AbstractVector
     },
     output_scitype = Union{
         Table(Union{Infinite, Finite}),
+        AbstractVector
     },
-    target_scitype = AbstractVector,
     load_path = "Imbalance.MLJ.SMOTENC"
 )
 
