+++ src/oversampling_methods/borderline_smote1/interface_mlj.jl
@@ -65,9 +65,20 @@ MMI.metadata_pkg(
 
 MMI.metadata_model(
     BorderlineSMOTE1,
-    input_scitype = Union{Table(Continuous),AbstractMatrix{Continuous}},
-    output_scitype = Union{Table(Continuous),AbstractMatrix{Continuous}},
-    target_scitype = AbstractVector,
+    input_scitype = Tuple{
+                        Union{
+                            Table(Continuous),
+                            AbstractMatrix{Continuous}
+                        }, 
+                        AbstractVector
+                    },
+    output_scitype = Tuple{
+        Union{
+            Table(Continuous),
+            AbstractMatrix{Continuous}
+        }, 
+        AbstractVector
+    },
     load_path = "Imbalance.MLJ.BorderlineSMOTE1"
 )
 function MMI.transform_scitype(s::BorderlineSMOTE1)
