+++ src/undersampling_methods/tomek_undersample/interface_mlj.jl
@@ -54,9 +54,20 @@ MMI.metadata_pkg(
 
 MMI.metadata_model(
     TomekUndersampler,
-    input_scitype = Union{Table(Continuous), AbstractMatrix{Continuous}},
-    output_scitype = Union{Table(Continuous), AbstractMatrix{Continuous}},
-    target_scitype = AbstractVector,
+    input_scitype = Tuple{
+                        Union{
+                            Table(Continuous),
+                            AbstractMatrix{Continuous}
+                        }, 
+                        AbstractVector
+                    },
+    output_scitype = Tuple{
+        Union{
+            Table(Continuous),
+            AbstractMatrix{Continuous}
+        }, 
+        AbstractVector
+    },
     load_path = "Imbalance.MLJ.TomekUndersampler"
 )
 function MMI.transform_scitype(s::TomekUndersampler)
