### 2024-01-09 b4c7ef9
🐛 Random Resampling should allow categorical types as well
