+++ src/undersampling_methods/random_undersample/interface_mlj.jl
@@ -52,8 +52,8 @@ MMI.metadata_model(
     RandomUndersampler,
     input_scitype = Tuple{
                         Union{
-                            Table(Continuous),
-                            AbstractMatrix{Continuous}
+                            Table(Union{Infinite, Finite}),
+                            AbstractMatrix{Infinite}
                         }, 
                         AbstractVector
                     },
