+++ src/oversampling_methods/smotenc/interface_mlj.jl
@@ -59,11 +59,11 @@ MMI.metadata_pkg(
 
 MMI.metadata_model(
     SMOTENC,
-    input_scitype = Union{
+    input_scitype = Tuple{
         Table(Union{Infinite, Finite}),
         AbstractVector
     },
-    output_scitype = Union{
+    output_scitype = Tuple{
         Table(Union{Infinite, Finite}),
         AbstractVector
     },
