+++ src/oversampling_methods/random_walk/interface_mlj.jl
@@ -44,11 +44,11 @@ MMI.metadata_pkg(
 
 MMI.metadata_model(
     RandomWalkOversampler,
-    input_scitype = Union{
+    input_scitype = Tuple{
         Table(Union{Infinite, Finite}),
         AbstractVector
     },
-    output_scitype = Union{
+    output_scitype = Tuple{
         Table(Union{Infinite, Finite}),
         AbstractVector
     },
