+++ src/oversampling_methods/smotenc/interface_mlj.jl
@@ -121,7 +121,11 @@ $((COMMON_DOCS["RNG"]))
 
 # Transform Inputs
 
-$((COMMON_DOCS["INPUTS"]))
+- `X`: A matrix of floats or a table with element [scitypes](https://juliaai.github.io/ScientificTypes.jl/) that subtype `Union{Finite, Infinite}`. 
+     Elements in nominal columns should subtype `Finite` (i.e., have [scitype](https://juliaai.github.io/ScientificTypes.jl/) `OrderedFactor` or `Multiclass`) and
+     elements in continuous columns should subtype `Infinite` (i.e., have [scitype](https://juliaai.github.io/ScientificTypes.jl/) `Count` or `Continuous`).
+
+- `y`: An abstract vector of labels (e.g., strings) that correspond to the observations in `X`
 
 # Transform Outputs
 
