### 2024-01-13 2ef8fb2
🚑 Allow matrices for SMOTENC and RandomWalk using MLJ
