### 2024-01-13 95d4d02
Revert "🚑 Allow matrices for SMOTENC and RandomWalk using MLJ"
This reverts commit 2ef8fb263eff382a2bf5ff90aa7d9d7916b58eeb.
