### 2024-01-13 179ee7f
🚑 Remove matrix support from MLJ docstring
