### 2024-01-21 #5971 @jbytecode
@EssamWisam - Hello, thank you for submitting your work to JOSS. I am the handling editor for this submission. The editorial bot has identified some previously published works that are similar to yours. Initially, I will attempt to invite the authors of these works to serve as reviewers. If we are unable to assign at least two reviewers through this process, I will then ask you to provide a list of potential reviewers from the suggested reviewers list. Thank you in advance.
