### 2024-01-21 #5971 @jbytecode
👋👋👋 Dear  @sylvaticus, @ArneTillmann, @matbesancon,  @MikeInnes 👋👋👋

**Imbalance: A comprehensive multi-interface Julia toolbox to address class imbalance**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5971).

Thank you in advance!
