### 2024-01-31 #6310 @editorialbot
[REVIEW]: Imbalance: A comprehensive multi-interface Julia toolbox to address class imbalance
**Submitting author:** @EssamWisam (<a href="http://orcid.org/0009-0009-1198-7166">Essam </a>)
**Repository:** https://github.com/JuliaAI/Imbalance.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.6
