### 2024-02-21 JuliaAI/Imbalance.jl#85 @sylvaticus
JOSS paper minor issues
Hello, great paper, just a few minor points for me:

a. Missing reference on line 31

b. "all four major types of resampling approaches" .. ok but which are these? From Table1 I could guess "Ensemble, Oversampling, Undersampling", but still missing one...

c. I think the sentence on BalanceModel/ MLJBalancing.jl comes too early.. I couldn't understood it from the paper, but from the readme in the package is ok.. I think you should skip it at that time, and eventually dedicate a few more lines after the design principles, well indicating that it isn't the MLJ interface for the package, that one is already in the Imbalance.jl package...

[part of https://github.com/openjournals/joss-reviews/issues/6310 ]
