### 2024-02-22 JuliaAI/Imbalance.jl#86 @sylvaticus
Documentation clarifications
Hello, I think the documentation is well done and very complete.
Just a few comments:

1. What I miss, as a naive user, is some guidance (just some sentence) on when to use one imbalance model or the other. Is there one imbalance model that is clearly superior nowadays ? Is producing sintetic data an advantage or a disadvantage compared to copy existing data ? What about oversampling vs undersampling ?

2. Are the TableInterface or MLJ interface different from model to model? If not, while it is important to have them documented, I think in the web site they make the documentation a bit redundant and difficult to follow.. I would just make a page for each of them (something like "MLJ Interface", "TableTransform Interface"), making just one full example and then saying they are available for [list of imbalance models], or perhaps list them with the options...

3. I haven't understand, for the `smote` function, how it is the default value for `k`, e.g. the sentence "It will be automatically set to m-1 for any class with m points where m ≤ k."

4. It's not clear to me what happens to the dataset when you use multiple balancers

5. In the "Combining Resamplers" page, I think you should remove the words "and in general" in the sentence "In prediction, the resamplers balancer1 and blancer2 are bypassed and in general."

I have tested all the algorithms with some continuous data and they seems to work.
I have also tested with some missing data in the X, but no algorithm works. I thought that at least the algorithms that simply randomly under/oversample should work with missing data, perhaps a note on this ?

Thank you for the great package!

[part of https://github.com/https://github.com/openjournals/joss-reviews/issues/6310 ]
