### 2024-02-22 JuliaAI/Imbalance.jl#85 @EssamWisam
Thank you for the comments!

For (a), will try to fix the missing reference, I just remember following the syntax and not having it work.

For (b), they are described in the Summary section of the paper:
`Two prevalent strategies for mitigating class imbalance.... It may be also possible to achieve even greater performance by combining both approaches or by resampling the data multiple times and ....
`
In other words, the fourth type is combination methods which are possible thanks to `MLJBalancing`.

I still completely agree that it should be further clarified. I will consider enumerating them in the same bullet point.

For (c), @ablaom  what do you think? I think our options are to phrase the following in a better way or push it down after the design principles?
`
Additionally, we present a companion package, MLJBalancing.jl, which: (i)  facilitates the integration of resampling methods with classification models via the BalancedModel construct, to create a seamless machine learning pipeline that behaves like a single unified model;  and (ii) implements a general version of the EasyEnsemble algorithm presented in `

CC: @ablaom
