### 2024-02-22 JuliaAI/Imbalance.jl#86 @EssamWisam
Thank you. Appreciate the feedback.

For (1), I remember deciding with @ablaom that the docs wouldn't be the best place for this which is why I created a series of four articles which are linked  in the docs (at the bottom of the main page) that aside from explaining the implemented algorithms give good hints for questions like `Is producing sintetic data an advantage or a disadvantage compared to copy existing data ?` and `What about oversampling vs undersampling ?` The question `Is there one imbalance model that is clearly superior nowadays ?` isn't  exactly explored in them but isn't easy to answer either as papers come from different times and evaluate on different datasets and because it's generally case-dependent. As in all machine learning, the user could always consider the choice of resampler itself as a hyperparameter to find what works best on their problem and I believe `MLJTuning.jl` may help in that.

Regarding (2), I absolutely agree that it could be viewed as redundant as these APIs are similar but didn't expect them to be as well difficult to follow. May in the future discuss and consider with @ablaom  making a single page for `MLJ` and another for`TableTransforms` where each explaisn how to infer the API given the pure functional interface and then lists the examples.

Regarding (3), the sentence `It will be automatically set to m-1 for any class with m points where m ≤ k.` In SMOTE, KNN is applied for data belonging to each class separately. Suppose `k` is chosen as 10 but one of the classes has 9 points then it will be set as `9-1=8` for that class as any point has 8 neighbors at most. I see now that not anyone using SMOTE knows this detail and will think if there is a better way to phrase it for that.

Regarding (4), a balancer is a mapping from `X, y` to `X',y'` where the data samples in `X, y,` are increased or decreased depending on whether it's oversampling or undersampling respectively. Consider `X` with three classes and a ratios of `1.0` this would oversample the data from all classes so that each of them has as much data as the majority class resulting in `X',y'` suppose we further use an undersampler such as `ENN` with `min_ratios=0.8` this would clean  the data around the decision boundaries to remove synthetic points in `X',y'` that likely would have made the classification task even harder while maintaining that no class falls bellow 80% of the number of points it had in `X',y'` resulting in the transformed dataset `X'',y''`. Does that make it clearer? I think having a clear of idea of what a single balancer does is enough to conclude what a chain would do.

Regarding (5), indeed, thanks.

> I have also tested with some missing data in the X, but no algorithm works. I thought that at least the algorithms that simply randomly under/oversample should work with missing data, perhaps a note on this ?

Fair point but deserves a separate issue optionally with a minimal reproducible example.

Thank you again for the great feedback!

CC @ablaom
