+++ README.md
@@ -1,0 +1,154 @@
+# Imbalance.jl
+
+![Imbalance](https://i.imgur.com/CP97JdN.png)
+
+
+A Julia package with resampling methods to correct for class imbalance in a wide variety of classification settings.
+
+
+
+[![Build Status](https://github.com/JuliaAI/Imbalance.jl/workflows/CI/badge.svg)](https://github.com/JuliaAI/Imbalance.jl/actions)
+[![codecov](https://codecov.io/gh/JuliaAI/Imbalance.jl/graph/badge.svg?token=K5UBZ3DCO6)](https://codecov.io/gh/JuliaAI/Imbalance.jl)
+[![Docs](https://img.shields.io/badge/docs-dev-blue.svg)](https://juliaai.github.io/Imbalance.jl/dev/)
+
+## ⏬ Installation
+```julia
+import Pkg;
+Pkg.add("Imbalance")
+```
+
+## ✨ Implemented Methods
+
+The package implements the following resampling algorithms
+
+- Random Oversampling
+- Random Walk Oversampling (RWO)
+- Random Oversampling Examples (ROSE)
+- Synthetic Minority Oversampling Technique (SMOTE)
+- Borderline SMOTE1
+- SMOTE-Nominal (SMOTE-N)
+- SMOTE-Nominal Categorical (SMOTE-NC)
+- Random Undersampling
+- Cluster Undersampling
+- EditedNearestNeighbors Undersampling
+- Tomek Links Undersampling
+- Balanced Bagging Classifier (@MLJBalancing.jl)
+
+To see various examples where such methods help improve classification performance, check the [tutorials sections](https://juliaai.github.io/Imbalance.jl/dev/examples/) of the documentation.
+
+Interested in contributing with more? Check [this](https://juliaai.github.io/Imbalance.jl/dev/contributing/).
+
+## 🚀 Quick Start
+
+We will illustrate using the package to oversample with`SMOTE`; however, all other implemented oversampling methods follow the same pattern.
+
+
+### 🔵 Standard API
+All methods by default support a pure functional interface.
+```julia
+using Imbalance
+
+# Set dataset properties then generate imbalanced data
+class_probs = [0.5, 0.2, 0.3]                  # probability of each class      
+num_rows, num_continuous_feats = 100, 5
+X, y = generate_imbalanced_data(num_rows, num_continuous_feats; class_probs, rng=42)      
+
+# Apply SMOTE to oversample the classes
+Xover, yover = smote(X, y; k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
+```
+In following code blocks, it will be assumed that `X` and `y` are readily available.
+
+### 🤖 MLJ Interface
+All methods support the [`MLJ` interface](https://alan-turing-institute.github.io/MLJ.jl/dev/) where instead of directly calling the method, one instantiates a model for the method while optionally passing the keyword parameters found in the functional interface then wraps the model in a `machine` and follows by calling `transform` on the machine and data.
+```julia
+using MLJ
+
+# Load the model
+SMOTE = @load SMOTE pkg=Imbalance
+
+# Create an instance of the model 
+oversampler = SMOTE(k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
+
+# Wrap it in a machine
+mach = machine(oversampler)
+
+# Provide the data to transform 
+Xover, yover = transform(mach, X, y)
+```
+All implemented oversampling methods are considered static transforms and hence, no `fit` is required. 
+
+#### Pipelining Models
[…]
