### 2024-03-01 #6310 @ArneTillmann
I am very happy with the quality of the submitted project and left my [comments](https://github.com/JuliaAI/Imbalance.jl/issues). As I am neither familiar with Julia, nor with the implementation of presented algorithms I rely somewhat on @sylvaticus review. If you, @jbytecode, are not satisfied with my review, please let me know as this is my first time :) Thank you very much
