+++ paper.md
@@ -48,7 +48,7 @@ Here `Xover, yover` are `X, y` after oversampling.
 
 A `ratios` hyperparameter or similar is always present to control the degree of oversampling or undersampling to be done for each class. All hyperparameters for a resampling method have default values that can be overridden.
 
-The set of resampling techniques implemented in `Imbalance.jl` and `MLJBalancing.jl` are shown in the table below. Note that although no combination resampling techniques are explicitly presented, they are easy to form using the `BalancedModel` wrapper found in `MLJBalancing.jl`.
+The set of resampling techniques implemented in either `Imbalance.jl` or `MLJBalancing.jl` are shown in the table below. Note that although no combination resampling techniques are explicitly presented, they are easy to form using the `BalancedModel` wrapper found in `MLJBalancing.jl` which can wrap an arbitrary number of resamplers in sequence.
 
 : Resampling techniques implemented in `Imbalance.jl` and `MLJBalancing.jl`. []{label="techniques"}
 
@@ -77,27 +77,27 @@ The toolbox implementation follows a specific set of design principles in terms 
 - Should offer solutions to heterogeneous data settings (continuous and nominal data)
 - When possible, preference should be given to techniques that are more common in the literature or industry
 
-Methods implemented in the `Imbalance.jl` toolbox indeed meet all aforementioned design principles for the implemented techniques. The one-vs-rest scheme as proposed in [@Fernández:2013] was used to generalize the technique to multi-class when needed.
+Methods implemented in the `Imbalance.jl` toolbox indeed meet all aforementioned design principles for the implemented techniques. The one-vs-rest scheme as proposed in [@Fernández:2013] was used to generalize binary technique to multi-class when needed.
 
 ### Interface Support
-- Should support both matrix and table inputs
+- Should support both matrix and table type inputs
 - Target variable may or may not be given as a separate column
 - Should expose a pure functional implementation, but also support popular Julia machine learning interfaces
-- Should be possible to wrap an arbitrary number of resampler models with an MLJ model to behave as a unified model using MLJBalancing
+- Should be possible to wrap an arbitrary number of resampler models with a classification model to behave as a unified model
 
 Methods implemented in the `Imbalance.jl` toolbox meet all the interface design principles above. It particularly implements the `MLJ` and `TableTransforms` interface for each method. `BalancedModel` from `MLJBalancing.jl` also allows fusing an arbitrary number of resampling models and a classifier together to behave as one unified model.
 
 
 ### Developer Experience and Testing
 
-- Should document all functions, including internal ones
-- Comments should be included to justify or simplify written implementations when needed
-- Features commonly used by multiple resampling techniques should be implemented in a single function and reused
+- There should exist a developer guide to encourage and guide contribution
 - Functions should be implemented in smaller units to aid in testing
 - Testing coverage should be maximized; even the most basic functions should be tested
-- There should exist a developer guide to encourage and guide contribution
+- Features commonly used by multiple resampling techniques should be implemented in a single function and reused
+- Should document all functions, including internal ones
+- Comments should be included to justify or simplify written implementations when needed
 
-This set of design principles is also satisfied by `Imbalance.jl`. Implemented techniques are tested by testing smaller units that form the technique. End-to-end tests are performed for each technique by testing properties and characteristics of the technique or by using the `imbalanced-learn` toolbox from Python and comparing outputs.
+This set of design principles is also satisfied by `Imbalance.jl`. Implemented techniques are tested by testing smaller units that form the technique. Aside from that, end-to-end tests are performed for each technique by testing properties and characteristics of the technique or by using the `imbalanced-learn` toolbox from Python and comparing outputs.
 
 ### User Experience
 
[…]
