+++ paper.bib
@@ -28,6 +28,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   title={Effective prediction of three common diseases by combining SMOTE with Tomek links technique for imbalanced medical data},
   author={Min Zeng and Beiji Zou and Faran Wei and Xiyao Liu and Lei Wang},
   journal={2016 IEEE International Conference of Online Analysis and Computing Science (ICOACS)},
+  doi={10.1109/ICOACS.2016.7563084},
   year={2016},
   pages={225-228},
   url={https://api.semanticscholar.org/CorpusID:25184489}
@@ -38,6 +39,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   title={Exploratory Undersampling for Class-Imbalance Learning},
   author={Xu-Ying Liu and Jianxin Wu and Zhi-Hua Zhou},
   journal={IEEE Transactions on Systems, Man, and Cybernetics, Part B (Cybernetics)},
+  doi={10.1109/TSMCB.2008.2007853},
   year={2009},
   volume={39},
   pages={539-550},
@@ -48,6 +50,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   title={The Curse of Class Imbalance and Conflicting Metrics with Machine Learning for Side-channel Evaluations},
   author={Stjepan Picek and Annelie Heuser and Alan Jovi{\'c} and Shivam Bhasin and Francesco Regazzoni},
   journal={IACR Trans. Cryptogr. Hardw. Embed. Syst.},
+  doi={10.13154/tches.v2019.i1.209-237},
   year={2018},
   volume={2019},
   pages={209-237},
@@ -66,6 +69,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
 @article{Chawla:2002,
   title={SMOTE: Synthetic Minority Over-sampling Technique},
   author={N. Chawla and K. Bowyer and Lawrence O. Hall and W. Philip Kegelmeyer},
+  doi={10.1613/jair.953},
   journal={ArXiv},
   year={2002},
   volume={abs/1106.1813},
@@ -77,6 +81,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   title={Borderline-SMOTE: A New Over-Sampling Method in Imbalanced Data Sets Learning},
   author={Hui Han and Wenyuan Wang and Binghuan Mao},
   booktitle={International Conference on Intelligent Computing},
+  doi={10.1007/11538059_91},
   year={2005},
   url={https://api.semanticscholar.org/CorpusID:12126950}
 }
@@ -87,6 +92,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   journal={Inf. Fusion},
   year={2014},
   volume={20},
+  doi={10.1016/j.inffus.2013.12.003},
   pages={99-116},
   url={https://api.semanticscholar.org/CorpusID:205432428}
 }
@@ -98,6 +104,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   journal={Data Mining and Knowledge Discovery},
   year={2012},
   volume={28},
+  doi={10.1007/s10618-012-0295-5},
   pages={92-122},
   url={https://api.semanticscholar.org/CorpusID:18164904}
 }
@@ -106,6 +113,7 @@ url="https://doi.org/10.1007/978-3-540-75171-7_2"
   title={Clustering-based undersampling in class-imbalanced data},
   author={Wei-Chao Lin and Chih-Fong Tsai and Ya-Han Hu and Jing-Shang Jhang},
   journal={Inf. Sci.},
+  doi={10.1016/j.ins.2017.05.008},
   year={2016},
   volume={409},
[…]
