### 2024-03-05 JuliaAI/Imbalance.jl#85 @EssamWisam
@sylvaticus I tried to improve all of (1), (2), (3) can you read the relevant sections of the paper and let me know if it's satisfactory now?

Thank you.
