### 2024-03-05 JuliaAI/Imbalance.jl#86 @EssamWisam
@sylvaticus I hope I appropriately responsed to all your questions/concerns. As an update, I improved the documentation for the third point (I stopped specifying the minor detail because it's covered in a warning internally anyway).
