### 2024-03-05 JuliaAI/Imbalance.jl#86 @sylvaticus
Hello, while re-reading the package doc, I discovered an imprecision.
On the page https://juliaai.github.io/Imbalance.jl/dev/examples/effect_of_s/effect_of_s/ you cite Decision Trees, then you describe what is essentially a kernel perceptron, and finally you employ in the example Bayesian linear discrimination analysis :-)
