### 2024-03-05 JuliaAI/Imbalance.jl#86 @EssamWisam
I see the error in the first one and I'm on the way to fix it. It's likely that I copied the previous notebook and modified it and missed this detail. Thanks for pointing that out.

I don't see the error in the second one. I say:
> Let's go for a decision tree from BetaML. We can't go for logistic regression as we did in the SMOTE tutorial because it does not support categotical features.

which is technically correct.
