### 2024-03-05 JuliaAI/Imbalance.jl#86 @sylvaticus
Yes, but a few lines earlier you says "Let's go for a logistic classifier form MLJLinearModels".

I think you need just to remove that sentence.
