### 2024-03-05 JuliaAI/Imbalance.jl#86 @sylvaticus
Here a MRE for the error when the X matrix contains  missing / non numerical values :

```julia

using Imbalance

y = ["A", "A", "B", "A", "B"]
X = [1 1.1 2.1;
     1 1.2 2.2;
     2 1.3 2.3;
     1 1.4 2.4;
     2 1.5 2.5; ]
Xover, yover = random_oversample(X, y) # ok

X = [1 1.1 2.1;
     1 1.2 2.2;
     2 1.3 2.3;
     1 1.4 missing;
     2 1.5 2.5; ]
Xover, yover = random_oversample(X, y) # error

X = ["a" 1.1 2.1;
     "a" 1.2 2.2;
     "b" 1.3 2.3;
     "a" 1.4 2.4;
     "b" 1.5 2.5; ]
Xover, yover = random_oversample(X, y) # error

X = ["a" "a";
     "a" "a";
     "b" "b";
     "a" "a";
     "b" "b"; ]
Xover, yover = random_oversample(X, y) # error
```

As I said, algorithms that simply randomly under/oversample without generating synthetic data shouldn't basically care anything about the eltype of the matrix, at least that's what users may expect. If the implementation is such that the matrix is converted in something and handling that cases are problematic independently of the algorithm, perhaps this would deserve a note in the documentation...
