### 2024-03-05 JuliaAI/Imbalance.jl#86 @EssamWisam
Thank you for the comment.

> perhaps this would deserve a note in the documentation...

The documentation so far suggests that matrix inputs should be of type `Real`. As far as I know, the type `Missing` (or type `String`) aren't instances of that. Likewise, it is also mentioned that tables can accept categorical inputs (so that's what should be used when the matrix has strings).

> As I said, algorithms that simply randomly under/oversample without generating synthetic data shouldn't basically care anything about the eltype of the matrix, at least that's what users may expect.

Irrespective of what I said above, I agree with this. I will make a PR soon so that when the input is a matrix, `random_oversample` and `random_undersample` accept real numbers as well as missing values.
