+++ src/oversampling_methods/random_oversample/random_oversample.jl
@@ -10,7 +10,7 @@ generate n new observations for that class using random oversampling
 - `Xnew`: A matrix where each column is a new observation generated by ROSE
 """
 function random_oversample_per_class(
-    X::AbstractMatrix{<:Real},
+    X::AbstractMatrix{<:Union{Real, Missing}},
     n::Integer;
     rng::AbstractRNG = default_rng(),
 )
@@ -129,12 +129,13 @@ A full basic example along with an animation can be found [here](https://githubt
     section which also explains running code on Google Colab.
 """
 function random_oversample(
-    X::AbstractMatrix{<:Real},
+    X::AbstractMatrix{<:Union{Real, Missing}},
     y::AbstractVector;
     ratios = 1.0,
     rng::Union{AbstractRNG,Integer} = default_rng(),
     try_preserve_type::Bool = true,
 )
+    println("lolss")
     rng = rng_handler(rng)
     Xover, yover = generic_oversample(X, y, random_oversample_per_class; ratios, rng,)
     return Xover, yover
