### 2024-03-05 JuliaAI/Imbalance.jl#86 @EssamWisam
Didn't mean to close it. Github did that by itself I guess.

In the PR above, I fixed the documentation for the two tutorials and made `random_oversample` and `random_undersample` accept missing values for matrix inputs (with tests).
