### 2024-03-06 JuliaAI/Imbalance.jl#86 @sylvaticus
Sorry, I didn't see the "A matrix of real numbers " in the doc.
Concerning the second typo in the documentation,  you ended up repeating the sentence  "Let's go for a decision tree from BetaML" . I made a small pull request to correct it, you can then close this issue, thank you.
