### 2024-03-13 f8266cd
fix bibtex and manuscript
