### 2024-03-13 #6310 @jbytecode
@EssamWisam - I have just sent a pull request: https://github.com/JuliaAI/Imbalance.jl/pull/96

This PR includes minor changes. Please review the PR. Please merge if you are okay with them. Thank you in advance.
