### 2024-03-13 #6310 @jbytecode
@EssamWisam - Could you please also check the tasks given in the previous post under the title **Additional Author Tasks After Review is Complete**? I need a version tag that covers the latest reviewed software and a Zenodo record registered with the create version task. In the Zenodo page, the title, the author list, their ORCIDs should match with those in the submission with the same exact order. Please report here the version tag, e.g. `v1.2.3` and the zenodo information here. Thank you in advance.
