### 2024-03-14 45e66a9
Revert "Check paper after reorder"
This reverts commit 81305036f3ea102b5e6d925edaad75bb162a3aec.
