### 2024-03-15 #6310 @jbytecode
@EssamWisam - Are you happy with the latest form of the pdf now? Table is splitted by the page break but it is not that bad to me.
