### 2024-03-15 #6310 @jbytecode
@EssamWisam - version registering stuff is independent of the paper check, so we can perform these task simultaneously.
