### 2024-03-15 #6310 @EssamWisam
Release number is 0.1.6 and this is the [release tag](https://github.com/JuliaAI/Imbalance.jl/releases/tag/v0.1.6).
