### 2024-03-17 ea01bc0
Replace two words and fix small typo
