+++ paper.bib
@@ -189,3 +189,14 @@ abstract = {Imbalanced classification problems are definitely around He and Grac
   pages={97-110},
   url={https://api.semanticscholar.org/CorpusID:131286}
 }
+
+@article{Blaom2020MLJAJ,
+  title={MLJ: A Julia package for composable Machine Learning},
+  author={Anthony D. Blaom and Franz J. Kir{\'a}ly and Thibaut Lienart and Yiannis Simillides and Diego Arenas and Sebastian J. Vollmer},
+  journal={J. Open Source Softw.},
+  doi={10.21105/joss.02704}
+  year={2020},
+  volume={5},
+  pages={2704},
+  url={https://api.semanticscholar.org/CorpusID:220768685}
+}
