+++ paper.bib
@@ -194,7 +194,7 @@ abstract = {Imbalanced classification problems are definitely around He and Grac
   title={MLJ: A Julia package for composable Machine Learning},
   author={Anthony D. Blaom and Franz J. Kir{\'a}ly and Thibaut Lienart and Yiannis Simillides and Diego Arenas and Sebastian J. Vollmer},
   journal={J. Open Source Softw.},
-  doi={10.21105/joss.02704}
+  doi={10.21105/joss.02704},
   year={2020},
   volume={5},
   pages={2704},
