### 2024-03-18 #6310 @EssamWisam
> @EssamWisam - I think we have nothing to do with those missing DOIs, right?

Yes, unlike other articles, I couldn't find that they have a doi.
